helix-flow/requirements.txt
aiohttp==3.8.5
apache-airflow==2.7.1
apache-airflow-providers-slack==8.1.0
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.7
pymongo==4.5.0
requests==2.31.0
slack-sdk==3.23.0
SQLAlchemy<2.0

